+++ Examples/Nodify.Playground/EditorSettings.cs
@@ -134,11 +134,18 @@
             set => SetProperty(ref _connectionSpacing, value);
         }
 
-        private ConnectionOffsetMode _connectionOffsetMode = ConnectionOffsetMode.None;
-        public ConnectionOffsetMode ConnectionOffsetMode
+        private ConnectionOffsetMode _srcConnectionOffsetMode = ConnectionOffsetMode.Static;
+        public ConnectionOffsetMode ConnectionSourceOffsetMode
         {
-            get => _connectionOffsetMode;
-            set => SetProperty(ref _connectionOffsetMode, value);
+            get => _srcConnectionOffsetMode;
+            set => SetProperty(ref _srcConnectionOffsetMode, value);
+        }
+
+        private ConnectionOffsetMode _targetConnectionOffsetMode = ConnectionOffsetMode.Static;
+        public ConnectionOffsetMode ConnectionTargetOffsetMode
+        {
+            get => _targetConnectionOffsetMode;
+            set => SetProperty(ref _targetConnectionOffsetMode, value);
         }
 
         private ArrowHeadEnds _arrowHeadEnds = ArrowHeadEnds.End;
@@ -148,28 +155,28 @@
             set => SetProperty(ref _arrowHeadEnds, value);
         }
 
-        private ArrowHeadShape _arrowHeadShape = ArrowHeadShape.Default;
+        private ArrowHeadShape _arrowHeadShape = ArrowHeadShape.Arrowhead;
         public ArrowHeadShape ArrowHeadShape
         {
             get => _arrowHeadShape;
             set => SetProperty(ref _arrowHeadShape, value);
         }
 
-        private PointEditor _connectionSourceOffset = new PointEditor { X = 15, Y = 0 };
+        private PointEditor _connectionSourceOffset = new PointEditor { X = 14, Y = 0 };
         public PointEditor ConnectionSourceOffset
         {
             get => _connectionSourceOffset;
             set => SetProperty(ref _connectionSourceOffset, value);
         }
 
-        private PointEditor _connectionTargetOffset = new PointEditor { X = 15, Y = 0 };
+        private PointEditor _connectionTargetOffset = new PointEditor { X = 14, Y = 0 };
         public PointEditor ConnectionTargetOffset
         {
             get => _connectionTargetOffset;
             set => SetProperty(ref _connectionTargetOffset, value);
         }
 
-        private PointEditor _connectionArrowSize = new PointEditor { X = 7, Y = 6 };
+        private PointEditor _connectionArrowSize = new PointEditor { X = 8, Y = 8 };
         public PointEditor ConnectionArrowSize
         {
             get => _connectionArrowSize;